+++ lumin/nn/callbacks/loss_callbacks.py
@@ -37,14 +37,13 @@ class GradClip(Callback):
     Arguments:
         clip: value to clip at
         clip_norm: whether to clip according to norm (`torch.nn.utils.clip_grad_norm_`) or value (`torch.nn.utils.clip_grad_value_`)
-        model: :class:`~lumin.nn.models.model.Model` with parameters to clip gradients, alternatively call :meth:`~lumin.nn.models.Model.set_model`
 
     Examples::
         >>> grad_clip = GradClip(1e-5)
     '''
 
-    def __init__(self, clip:float, clip_norm:bool=True, model:Optional[AbsModel]=None):
-        super().__init__(model=model)
+    def __init__(self, clip:float, clip_norm:bool=True):
+        super().__init__()
         self.clip = clip
         self.func = nn.utils.clip_grad_norm_ if clip_norm else nn.utils.clip_grad_value_